Minor bugfix for Visual Studio 2017.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@4149 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/pr_list_head.c
@@ -1020,7 +1020,7 @@ mx_list_head_record_show_socket_handlers( MX_LIST_HEAD *list_head )
 		  "socket_handler(%d) = %p, socket id = %d,\n"
 		  "        username = '%s', program = '%s', process_id = %lu\n",
 				i, socket_handler,
-				socket_handler->synchronous_socket->socket_fd,
+			(int) socket_handler->synchronous_socket->socket_fd,
 				socket_handler->username,
 				socket_handler->program_name,
 				socket_handler->process_id );
